# backend/src/main/java/com/busticket/app/model/dto/response/PlacanjeResponseDTO.java
package com.busticket.app.model.dto.response;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class PlacanjeResponseDTO {

    private Long id;
    private String stripePaymentId;
    private double iznos;
    private LocalDateTime datum;
    private String status;
    private Long rezervacijaId;

}