# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Bus Ticket App
  description: |-
    REST API for managing bus ticket reservations, trips, companies, vehicles and payments.
    Built with Spring Boot and integrated with Stripe for payment processing.

    Some useful links:
    - [Bus Ticket App repository](https://github.com/texksi/Bus-Ticket-App)
  contact:
    email: teamikic18@gmail.com
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/placanja/create-payment:
    post:
      tags:
        - Placanje
      summary: Create Stripe payment intent
      parameters:
        - name: rezervacijaId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: iznos
          in: query
          required: true
          schema:
            type: number
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlacanjeResponse'
...
  /api/vozila:
    post:
      tags:
        - Vozilo
      summary: Create new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoziloRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoziloResponse'
        '409':
          description: Vehicle with this registration already exists
  /api/vozila/{id}:
    get:
      tags:
        - Vozilo
      summary: Get vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoziloResponse'
        '404':
          description: Vehicle not found
    put:
      tags:
        - Vozilo
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoziloRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoziloResponse'
    delete:
      tags:
        - Vozilo
      summary: Delete vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /api/vozila/kompanija/{id}:
    get:
      tags:
        - Vozilo
      summary: Get all vehicles for company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoziloResponse'
components:
  schemas:
...
    PlacanjeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        stripePaymentId:
          type: string
        iznos:
          type: number
        datum:
          type: string
          format: date-time
        status:
          type: string
        rezervacijaId:
          type: integer
          format: int64
...
    VoziloRequest:
      type: object
      properties:
        registracija:
          type: string
          example: "BG123AB"
        kapacitet:
          type: integer
          example: 50
        brojRedova:
          type: integer
          example: 10
        brojKolona:
          type: integer
          example: 5
        kompanijaId:
          type: integer
          format: int64
          example: 1
    VoziloResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        registracija:
          type: string
        kapacitet:
          type: integer
        brojRedova:
          type: integer
        brojKolona:
          type: integer
        kompanijaId:
          type: integer
          format: int64